timer.py changes (BROKEN)
Added input box for inputting time in seconds.

Also added start button and timing functions

currently doesn't work because it calls update_timer() at init but sees that time_left is 0 so then does nothing. I want to add time with start button when timer is running and start the timer when the timer isn't running.